requirements.txt
amqp==5.3.1
aniso8601==7.0.0
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
cffi==1.17.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.6
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-graphql-jwt==0.4.0
djangorestframework==3.15.2
drf-yasg==1.21.9
flake8==7.1.2
graphene==2.1.9
graphene-django==2.16.0
graphql-core==2.3.2
graphql-relay==2.0.1
inflection==0.5.1
kombu==5.4.2
mccabe==0.7.0
packaging==24.2
promise==2.3
prompt_toolkit==3.0.50
psycopg==3.2.5
psycopg-binary==3.2.5
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
pytz==2025.1
PyYAML==6.0.2
rabbitmq==0.2.0
Rx==1.6.3
singledispatch==4.1.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
vine==5.1.0
wcwidth==0.2.13
